+++ kroxylicious-sample/src/test/java/io/kroxylicious/sample/SampleFilterIT.java
@@ -47,9 +47,6 @@ class SampleFilterIT {
     private static final TestFilter SAMPLE_FETCH_RESPONSE_FILTER = new TestFilter(SampleFetchResponse.class.getName(),
             Map.of(FIND_CONFIG_FIELD, "bar", REPLACE_CONFIG_FIELD, "baz"));
 
-    private static final TestFilter SAMPLE_FETCH_RESPONSE_FILTER_DEPRECATED_FACTORY_NAME = new TestFilter(SampleFetchResponseFilterFactory.class.getName(),
-            Map.of(FIND_CONFIG_FIELD, "bar", REPLACE_CONFIG_FIELD, "baz"));
-
     // Configure test input/expected values here
     private static final String NO_TRANSFORM_VALUE = "sample";
     private static final String PRE_TRANSFORM_VALUE = "foo bar baz";
@@ -112,14 +109,6 @@ class SampleFilterIT {
     @Test
     void sampleFetchResponseFilterWillTransform() {
         test = new FilterIntegrationTest(SAMPLE_FETCH_RESPONSE_FILTER);
-        test.produceMessage(PRE_TRANSFORM_VALUE)
-                .consumeSingleRecord()
-                .assertConsumerRecordEquals(FETCH_TRANSFORM_VALUE);
-    }
-
-    @Test
-    void deprecatedSampleFetchResponseFilterWillTransform() {
-        test = new FilterIntegrationTest(SAMPLE_FETCH_RESPONSE_FILTER_DEPRECATED_FACTORY_NAME);
         test.produceMessage(PRE_TRANSFORM_VALUE)
                 .consumeSingleRecord()
                 .assertConsumerRecordEquals(FETCH_TRANSFORM_VALUE);